feat(frontend): Add management command to run backend tests
I have overridden the `./manage.py test` command with a custom one,
to use `pytest` and `ptw` commands. We exclusively use pytest, so
no point having the existing `test` command. The new command allows
us to run tests in `--single` mode, but the default is in `watch` mode.
This basically re-runs tests when changes are detected.

IssueID #